+++ cartographer/internal/mapping/global_trajectory_builder.h
@@ -20,110 +20,17 @@
 #include "cartographer/mapping/trajectory_builder_interface.h"
 
 #include "cartographer/mapping/local_slam_result_data.h"
-#include "glog/logging.h"
 
 namespace cartographer {
 namespace mapping {
 
 template <typename LocalTrajectoryBuilder,
           typename LocalTrajectoryBuilderOptions, typename PoseGraph>
-class GlobalTrajectoryBuilder : public mapping::TrajectoryBuilderInterface {
- public:
-  // Passing a 'nullptr' for 'local_trajectory_builder' is acceptable, but no
-  // 'TimedPointCloudData' may be added in that case.
-  GlobalTrajectoryBuilder(
-      std::unique_ptr<LocalTrajectoryBuilder> local_trajectory_builder,
-      const int trajectory_id, PoseGraph* const pose_graph,
-      const LocalSlamResultCallback& local_slam_result_callback)
-      : trajectory_id_(trajectory_id),
-        pose_graph_(pose_graph),
-        local_trajectory_builder_(std::move(local_trajectory_builder)),
-        local_slam_result_callback_(local_slam_result_callback) {}
-  ~GlobalTrajectoryBuilder() override {}
-
-  GlobalTrajectoryBuilder(const GlobalTrajectoryBuilder&) = delete;
-  GlobalTrajectoryBuilder& operator=(const GlobalTrajectoryBuilder&) = delete;
-
-  void AddSensorData(
-      const std::string& sensor_id,
-      const sensor::TimedPointCloudData& timed_point_cloud_data) override {
-    CHECK(local_trajectory_builder_)
-        << "Cannot add TimedPointCloudData without a LocalTrajectoryBuilder.";
-    std::unique_ptr<typename LocalTrajectoryBuilder::MatchingResult>
-        matching_result = local_trajectory_builder_->AddRangeData(
-            timed_point_cloud_data.time,
-            sensor::TimedRangeData{timed_point_cloud_data.origin,
-                                   timed_point_cloud_data.ranges,
-                                   {}});
-    if (matching_result == nullptr) {
-      // The range data has not been fully accumulated yet.
-      return;
-    }
-    std::unique_ptr<InsertionResult> insertion_result;
-    if (matching_result->insertion_result != nullptr) {
-      auto node_id = pose_graph_->AddNode(
-          matching_result->insertion_result->constant_data, trajectory_id_,
-          matching_result->insertion_result->insertion_submaps);
-      CHECK_EQ(node_id.trajectory_id, trajectory_id_);
-      insertion_result = common::make_unique<InsertionResult>(InsertionResult{
-          node_id, matching_result->insertion_result->constant_data,
-          std::vector<std::shared_ptr<const Submap>>(
-              matching_result->insertion_result->insertion_submaps.begin(),
-              matching_result->insertion_result->insertion_submaps.end())});
-    }
-    if (local_slam_result_callback_) {
-      local_slam_result_callback_(
-          trajectory_id_, matching_result->time, matching_result->local_pose,
-          std::move(matching_result->range_data_in_local),
-          std::move(insertion_result));
-    }
-  }
-
-  void AddSensorData(const std::string& sensor_id,
-                     const sensor::ImuData& imu_data) override {
-    if (local_trajectory_builder_) {
-      local_trajectory_builder_->AddImuData(imu_data);
-    }
-    pose_graph_->AddImuData(trajectory_id_, imu_data);
-  }
-
-  void AddSensorData(const std::string& sensor_id,
-                     const sensor::OdometryData& odometry_data) override {
-    CHECK(odometry_data.pose.IsValid()) << odometry_data.pose;
-    if (local_trajectory_builder_) {
-      local_trajectory_builder_->AddOdometryData(odometry_data);
-    }
-    pose_graph_->AddOdometryData(trajectory_id_, odometry_data);
-  }
-
-  void AddSensorData(
-      const std::string& sensor_id,
-      const sensor::FixedFramePoseData& fixed_frame_pose) override {
-    if (fixed_frame_pose.pose.has_value()) {
-      CHECK(fixed_frame_pose.pose.value().IsValid())
-          << fixed_frame_pose.pose.value();
-    }
-    pose_graph_->AddFixedFramePoseData(trajectory_id_, fixed_frame_pose);
-  }
-
-  void AddSensorData(const std::string& sensor_id,
-                     const sensor::LandmarkData& landmark_data) override {
-    pose_graph_->AddLandmarkData(trajectory_id_, landmark_data);
-  }
-
-  void AddLocalSlamResultData(std::unique_ptr<mapping::LocalSlamResultData>
-                                  local_slam_result_data) override {
-    CHECK(!local_trajectory_builder_) << "Can't add LocalSlamResultData with "
-                                         "local_trajectory_builder_ present.";
-    local_slam_result_data->AddToPoseGraph(trajectory_id_, pose_graph_);
-  }
-
- private:
-  const int trajectory_id_;
-  PoseGraph* const pose_graph_;
-  std::unique_ptr<LocalTrajectoryBuilder> local_trajectory_builder_;
-  LocalSlamResultCallback local_slam_result_callback_;
-};
+std::unique_ptr<TrajectoryBuilderInterface> CreateGlobalTrajectoryBuilder(
+    std::unique_ptr<LocalTrajectoryBuilder> local_trajectory_builder,
+    const int trajectory_id, PoseGraph* const pose_graph,
+    const TrajectoryBuilderInterface::LocalSlamResultCallback&
+        local_slam_result_callback);
 
 }  // namespace mapping
 }  // namespace cartographer
